clear routes
