+++ server/index.js
@@ -9,10 +9,14 @@ var app = express();
 
 app.use(express.static(path.join(config.root, 'static')));
 
-require('./components').init(app, function () {
-    app.listen(config.port, config.ip, function () {
-        console.info('Express server listening on %d, in %s mode', config.port, config.env)
-    })
+require('./components').init(app, function (err) {
+
+    if (err)
+        console.log(err);
+    else
+        app.listen(config.port, config.ip, function () {
+            console.info('Express server listening on %d, in %s mode', config.port, config.env)
+        })
 });
